+++ app/Http/Controllers/KnowledgeController.php
@@ -2,7 +2,7 @@
 
 namespace App\Http\Controllers;
 
-use App\Models\Knowledge;
+use App\Models\{Knowledge, User};
 use Illuminate\Http\Request;
 use App\Service\KnowledgeService;
 
@@ -21,8 +21,13 @@ class KnowledgeController extends Controller
      */
     public function index()
     {
-        $knowledges = $this->knowledgeService->getKnowledges();
-        return view("knowledges.index", ['knowls' => $knowledges]);
+        if(\Auth::check()){
+            $this->authorize(User::VIEW_ANY, Knowledge::class);
+            $knowledges = $this->knowledgeService->getKnowledges();
+            return view("knowledges.index", ['knowls' => $knowledges]);
+        } else {
+            return redirect()->route('user.login');
+        }
     }
 
     /**
@@ -32,7 +37,12 @@ class KnowledgeController extends Controller
      */
     public function create()
     {
-        return view("knowledges.create");
+        if(\Auth::check()){
+            $this->authorize(User::CREATE, Knowledge::class);
+            return view("knowledges.create");
+        } else {
+            return redirect()->route('user.login');
+        }
     }
 
     /**
@@ -43,16 +53,21 @@ class KnowledgeController extends Controller
      */
     public function store(Request $request)
     {
-        $this->validate($request, [
-            'user_id' => 'required',
-            'name' => 'required',
-            'description' => 'required',
-            "data" => 'required'
-        ]);
+        if(\Auth::check()){
+            $this->authorize(User::CREATE, Knowledge::class);
+            $this->validate($request, [
+                'user_id' => 'required',
+                'name' => 'required',
+                'description' => 'required',
+                "data" => 'required'
+            ]);
 
-        $data = $request->only(['name', 'description', 'data', 'user_id']);
-        $this->knowledgeService->createKnowledge($data);
-        return redirect()->back();
+            $data = $request->only(['name', 'description', 'data', 'user_id']);
+            $this->knowledgeService->createKnowledge($data);
+             return redirect()->back();
+        } else {
+            return redirect()->route('user.login');
+        }
     }
 
     /**
@@ -63,8 +78,13 @@ class KnowledgeController extends Controller
      */
     public function show($id)
     {
-        $knowl = $this->knowledgeService->getKnowledge($id);
-        return view("knowledges.show", ["knowl" => $knowl]);
+        if(\Auth::check()){
+            $knowl = $this->knowledgeService->getKnowledge($id);
+            $this->authorize(User::VIEW, $knowl);
+            return view("knowledges.show", ["knowl" => $knowl]);
+        } else {
+            return redirect()->route('user.login');
+        }
     }
 
     /**
@@ -75,8 +95,13 @@ class KnowledgeController extends Controller
      */
     public function edit($id)
     {
-        $knowl = $this->KnowledgeService->getKnowledge($id);
-        return view("knowledges.edit", ["knowl" => $knowl]);
+        if(\Auth::check()){
+            $knowl = $this->knowledgeService->getKnowledge($id);
+            $this->authorize(User::UPDATE, $knowl);
+            return view("knowledges.edit", ["knowl" => $knowl]);
+        } else {
+            return redirect()->route('user.login');
+        }
     }
 
     /**
@@ -88,16 +113,22 @@ class KnowledgeController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $this->validate($request, [
-            'user_id' => 'required',
-            'name' => 'required',
-            'description' => 'required',
-            "data" => 'required'
-        ]);
+        if(\Auth::check()){
+            $this->validate($request, [
+                'user_id' => 'required',
+                'name' => 'required',
+                'description' => 'required',
+                "data" => 'required'
+            ]);
 
-        $data = $request->only(['name', 'description', 'data', 'user_id']);
-        $this->knowledgeService->updateKnowledge($data, $id);
-        return redirect()->back();
+            $data = $request->only(['name', 'description', 'data', 'user_id']);
+            $updatingKnowledge = $this->knowledgeService->getKnowledge($id);
+            $this->authorize(User::UPDATE, $updatingKnowledge);
+            $this->knowledgeService->updateKnowledge($data, $id);
+            return redirect()->back();
+        } else {
+            return redirect()->route('user.login');
+        }
     }
 
     /**
@@ -108,7 +139,13 @@ class KnowledgeController extends Controller
      */
     public function destroy($id)
     {
-        $this->knowledgeService->deleteKnowledge($id);
-        return redirect()->back();
+        if(Auth::check()){
+            $updatingKnowledge = $this->knowledgeService->getKnowledge($id);
+            $this->authorize(User::DELETE, $updatingKnowledge);
+            $this->knowledgeService->deleteKnowledge($id);
+            return redirect()->back();
+        } else {
+            return redirect()->route('user.login');
+        }
     }
 }
