+++ app/Http/Controllers/TaskController.php
@@ -2,7 +2,7 @@
 
 namespace App\Http\Controllers;
 
-use App\Models\Task;
+use App\Models\{Task, User};
 use Illuminate\Http\Request;
 use App\Service\TaskService;
 
@@ -22,8 +22,13 @@ class TaskController extends Controller
      */
     public function index()
     {
-        $tasks = $this->taskService->getTasks();
-        return view("tasks.index", ["tasks" => $tasks]);
+        if(\Auth::check()){
+            $this->authorize(User::VIEW_ANY, Task::class);
+            $tasks = $this->taskService->getTasks();
+            return view("tasks.index", ["tasks" => $tasks]);
+        } else {
+            return redirect()->route('user.login');
+        }
     }
 
     /**
@@ -33,7 +38,11 @@ class TaskController extends Controller
      */
     public function create()
     {
-        return view("tasks.create");
+        if(\Auth::check()){
+            return view("tasks.create");
+        } else {
+            return redirect()->route('user.login');
+        }
     }
 
     /**
@@ -44,14 +53,18 @@ class TaskController extends Controller
      */
     public function store(Request $request)
     {
-        $this->validate($request, [
-            'name' => 'required|max:255',
-        ]);
+        if(\Auth::check()){
+            $this->validate($request, [
+                'name' => 'required|max:255',
+            ]);
 
-        // create record and pass in only fields that are fillable
-        $data = $request->only(["name", "description", "status"]);
-        $this->taskService->createTask($data);
-        return redirect()->back();
+            // create record and pass in only fields that are fillable
+            $data = $request->only(["name", "description", "status"]);
+            $this->taskService->createTask($data);
+            return redirect()->back();
+        } else{
+            return redirect()->route('user.login');
+        }
     }
 
     /**
@@ -62,8 +75,13 @@ class TaskController extends Controller
      */
     public function show($id)
     {
-        $task = $this->taskService->getTask($id);
-        return view("tasks.show", ["task" => $task]);
+        if(\Auth::check()){
+            $task = $this->taskService->getTask($id);
+            $this->authorize(User::VIEW, $task);
+            return view("tasks.show", ["task" => $task]);
+        } else {
+            return redirect()->route('user.login');
+        }
     }
 
     /**
@@ -74,8 +92,13 @@ class TaskController extends Controller
      */
     public function edit($id)
     {
-        $task = $this->taskService->getTask($id);
-        return view("tasks.edit", ["task" => $task]);
+        if(\Auth::check()){
+            $authUser = \Auth::user();
+            $task = $this->taskService->getTask($id);
+            if($authUser->can(User::UPDATE, $task)){
+                return view("tasks.edit", ["task" => $task]);
+            }
+        }
     }
 
     /**
@@ -87,14 +110,19 @@ class TaskController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $this->validate($request, [
-            'name' => 'required|max:255',
-        ]);
+        if(\Auth::check()){
+            $this->validate($request, [
+                'name' => 'required|max:255',
+            ]);
 
-        $data = $request->only(["name", "description", "status"]);
-        $this->taskService->updateTask($data, $id);
-
-        return redirect()->back();
+            $data = $request->only(["name", "description", "status"]);
+            $updatingTask = $this->taskService->getTask($id);
+            $this->authorize(User::UPDATE, $updatingTask);
+            $this->taskService->updateTask($data, $id);
+            return redirect()->back();
+        } else {
+            return redirect()->route('user.login');
+        }
     }
 
     /**
@@ -105,7 +133,13 @@ class TaskController extends Controller
      */
     public function destroy($id)
     {
-        $this->taskService->deleteTask($id);
-        return redirect()->back();
+        if(Auth::check()){
+            $updatingTask = $this->taskService->getTask($id);
+            $this->authorize(User::DELETE, $updatingTask);
+            $this->taskService->deleteTask($id);
+            return redirect()->back();
+        } else {
+            return redirect()->route('user.login');
+        }
     }
 }
